Module rustc_trans::back::link

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Constants

ALLOCATOR_MODULE_NAME 	 [
Experimental
]

ALLOCATOR_OBJ_NAME 	 [
Experimental
]

METADATA_MODULE_NAME 	 [
Experimental
]
The LLVM module name containing crate-metadata. This includes a . on purpose, so it cannot clash with the name of a user-defined module.

METADATA_OBJ_NAME 	 [
Experimental
]
The name of the crate-metadata object file the compiler generates. Must match up with METADATA_MODULE_NAME.

RLIB_BYTECODE_OBJECT_MAGIC 	 [
Experimental
]

RLIB_BYTECODE_OBJECT_V1_DATASIZE_OFFSET 	 [
Experimental
]

RLIB_BYTECODE_OBJECT_V1_DATA_OFFSET 	 [
Experimental
]

RLIB_BYTECODE_OBJECT_VERSION 	 [
Experimental
]

RLIB_BYTECODE_OBJECT_VERSION_OFFSET 	 [
Experimental
]

Functions

build_link_meta 	 [
Experimental
]

default_output_for_target 	 [
Experimental
]
Returns default crate type for target

each_linked_rlib 	 [
Experimental
]

filename_for_input 	 [
Experimental
]

find_crate_name 	 [
Experimental
]

get_linker 	 [
Experimental
]

ignored_for_lto 	 [
Experimental
]
Returns a boolean indicating whether the specified crate should be ignored during LTO.

invalid_output_for_target 	 [
Experimental
]
Checks if target supports crate_type as output

link_binary 	 [
Experimental
]
Perform the linkage portion of the compilation phase. This will generate all of the requested outputs for this compilation session.

msvc_link_exe_cmd 	 [
Experimental
]

remove 	 [
Experimental
]