Function rustc_trans::back::link::ignored_for_lto

pub fn ignored_for_lto(info: &CrateInfo, cnum: CrateNum) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns a boolean indicating whether the specified crate should be ignored during LTO.

Crates ignored during LTO are not lumped together in the "massive object file" that we create and are linked in their normal rlib states. See comments below for what crates do not participate in LTO.

It's unusual for a crate to not participate in LTO. Typically only compiler-specific and unstable crates have a reason to not participate in LTO.