Module rustc_trans::back::msvc

Unstable (rustc_private)

MSVC-specific logic for linkers and such.

This module contains a cross-platform interface but has a blank unix implementation. The Windows implementation builds on top of Windows native libraries (reading registry keys), so it otherwise wouldn't link on unix.

Note that we don't have much special logic for finding the system linker on any other platforms, so it may seem a little odd to single out MSVC to have a good deal of code just to find the linker. Unlike Unix systems, however, the MSVC linker is not in the system PATH by default. It also additionally needs a few environment variables or command line flags to be able to link against system libraries.

In order to have a nice smooth experience on Windows, the logic in this file is here to find the MSVC linker and set it up in the default configuration one would need to set up anyway. This means that the Rust compiler can be run not only in the developer shells of MSVC but also the standard cmd.exe shell or MSYS shells.

As a high-level note, all logic in this module for looking up various paths/files is based on Microsoft's logic in their vcvars bat files, but comments can also be found below leading through the various code paths.

Functions

host_dll_path 	 [Unstable]

link_exe_cmd 	 [Unstable]