Module rustc_trans::back::write

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

CodegenContext 	 [
Experimental
]
Additional resources used by optimize_and_codegen (not module specific)

CompiledModules 	 [
Experimental
]

ModuleConfig 	 [
Experimental
]
Module-specific configuration for optimize_and_codegen.

OngoingCrateTranslation 	 [
Experimental
]

SharedEmitter 	 [
Experimental
]

SharedEmitterMain 	 [
Experimental
]

Constants

CODE_GEN_MODEL_ARGS 	 [
Experimental
]

RELOC_MODEL_ARGS 	 [
Experimental
]

TRANS_WORKER_ID 	 [
Experimental
]

TRANS_WORKER_TIMELINE 	 [
Experimental
]

TRANS_WORK_PACKAGE_KIND 	 [
Experimental
]

Functions

create_target_machine 	 [
Experimental
]

dump_incremental_data 	 [
Experimental
]

llvm_err 	 [
Experimental
]

run_assembler 	 [
Experimental
]

start_async_translation 	 [
Experimental
]

with_llvm_pmb⚠ 	 [
Experimental
]

write_output_file 	 [
Experimental
]