Module rustc_trans::lint::builtin

Unstable (rustc_private)

Some lints that are built in to the compiler.

These are the built-in lints that are emitted direct in the main compiler code, rather than using their own custom pass. Those lints are all available in rustc_lint::builtin.

Structs

HardwiredLints 	 [Unstable]
Does nothing as a lint pass, but registers some Lints which are used by other parts of the compiler.

Statics

CONST_ERR 	 [Unstable]

DEAD_CODE 	 [Unstable]

FAT_PTR_TRANSMUTES 	 [Unstable]

HR_LIFETIME_IN_ASSOC_TYPE 	 [Unstable]

ILLEGAL_FLOATING_POINT_CONSTANT_PATTERN 	 [Unstable]

ILLEGAL_STRUCT_OR_ENUM_CONSTANT_PATTERN 	 [Unstable]

INACCESSIBLE_EXTERN_CRATE 	 [Unstable]

INVALID_TYPE_PARAM_DEFAULT 	 [Unstable]

MATCH_OF_UNIT_VARIANT_VIA_PAREN_DOTDOT 	 [Unstable]

OBJECT_UNSAFE_FRAGMENT 	 [Unstable]

OVERLAPPING_INHERENT_IMPLS 	 [Unstable]

PRIVATE_IN_PUBLIC 	 [Unstable]

RAW_POINTER_DERIVE 	 [Unstable]

RENAMED_AND_REMOVED_LINTS 	 [Unstable]

STABLE_FEATURES 	 [Unstable]

SUPER_OR_SELF_IN_GLOBAL_PATH 	 [Unstable]

TRANSMUTE_FROM_FN_ITEM_TYPES 	 [Unstable]

TRIVIAL_CASTS 	 [Unstable]

TRIVIAL_NUMERIC_CASTS 	 [Unstable]

UNKNOWN_CRATE_TYPES 	 [Unstable]

UNKNOWN_LINTS 	 [Unstable]

UNREACHABLE_CODE 	 [Unstable]

UNSIZED_IN_TUPLE 	 [Unstable]

UNUSED_ASSIGNMENTS 	 [Unstable]

UNUSED_EXTERN_CRATES 	 [Unstable]

UNUSED_FEATURES 	 [Unstable]

UNUSED_IMPORTS 	 [Unstable]

UNUSED_QUALIFICATIONS 	 [Unstable]

UNUSED_VARIABLES 	 [Unstable]

VARIANT_SIZE_DIFFERENCES 	 [Unstable]

WARNINGS 	 [Unstable]