Module rustc_trans::metadata

Unstable

Modules

common 	 [Unstable]

creader 	 [Unstable]
Validates all used crates and extern libraries and loads their metadata

csearch 	 [Unstable]

cstore 	 [Unstable]

decoder 	 [Unstable]

encoder 	 [Unstable]

filesearch 	 [Unstable]

loader 	 [Unstable]
Finds crate binaries and loads their metadata

macro_import 	 [Unstable]
Used by rustc when loading a crate with exported macros.

tydecode 	 [Unstable]

tyencode 	 [Unstable]