Module rustc_trans::metadata::creader

Unstable

Validates all used crates and extern libraries and loads their metadata

Structs

CrateReader 	 [Unstable]

Functions

import_codemap 	 [Unstable]
Imports the codemap from an external crate into the codemap of the crate currently being compiled (the "local crate").

validate_crate_name 	 [Unstable]