Module rustc_trans::metadata::decoder

Unstable (rustc_private)

Structs

CrateDep 	 [Unstable]

Enums

DefLike 	 [Unstable]

Functions

closure_kind 	 [Unstable]

closure_ty 	 [Unstable]

crate_rustc_version 	 [Unstable]

def_path 	 [Unstable]

each_child_of_item 	 [Unstable]
Iterates over each child of the given item.

each_exported_macro 	 [Unstable]

each_implementation_for_trait 	 [Unstable]

each_inherent_implementation_for_type 	 [Unstable]

each_lang_item 	 [Unstable]
Iterates over the language items in the given crate.

each_top_level_item_of_crate 	 [Unstable]
Iterates over all the top-level crate items.

get_adt_def 	 [Unstable]

get_associated_consts 	 [Unstable]

get_crate_attributes 	 [Unstable]

get_crate_deps 	 [Unstable]

get_crate_hash 	 [Unstable]

get_crate_name 	 [Unstable]

get_crate_triple 	 [Unstable]

get_custom_coerce_unsized_kind 	 [Unstable]

get_dylib_dependency_formats 	 [Unstable]

get_impl_items 	 [Unstable]
Returns the def IDs of all the items in the given implementation.

get_impl_or_trait_item 	 [Unstable]

get_impl_polarity 	 [Unstable]

get_impl_trait 	 [Unstable]

get_imported_filemaps 	 [Unstable]

get_item_attrs 	 [Unstable]

get_item_name 	 [Unstable]

get_item_path 	 [Unstable]

get_item_variances 	 [Unstable]

get_method_arg_names 	 [Unstable]

get_methods_if_impl 	 [Unstable]

get_missing_lang_items 	 [Unstable]

get_native_libraries 	 [Unstable]

get_plugin_registrar_fn 	 [Unstable]

get_predicates 	 [Unstable]

get_provided_trait_methods 	 [Unstable]

get_reachable_ids 	 [Unstable]

get_repr_attrs 	 [Unstable]

get_stability 	 [Unstable]

get_struct_field_attrs 	 [Unstable]

get_struct_field_names 	 [Unstable]

get_super_predicates 	 [Unstable]

get_symbol 	 [Unstable]

get_symbol_from_buf 	 [Unstable]
If you have a crate_metadata, call get_symbol instead

get_trait_def 	 [Unstable]

get_trait_item_def_ids 	 [Unstable]

get_trait_name 	 [Unstable]

get_trait_of_item 	 [Unstable]

get_tuple_struct_definition_if_ctor 	 [Unstable]
If node_id is the constructor of a tuple struct, retrieve the NodeId of the actual type definition, otherwise, return None

get_type 	 [Unstable]

is_const_fn 	 [Unstable]

is_default_impl 	 [Unstable]

is_defaulted_trait 	 [Unstable]

is_extern_fn 	 [Unstable]

is_impl 	 [Unstable]

is_static 	 [Unstable]

is_static_method 	 [Unstable]

is_typedef 	 [Unstable]

item_type 	 [Unstable]

list_crate_metadata 	 [Unstable]

load_index 	 [Unstable]

load_xrefs 	 [Unstable]

maybe_get_crate_hash 	 [Unstable]

maybe_get_crate_name 	 [Unstable]

maybe_get_item_ast 	 [Unstable]

translate_def_id 	 [Unstable]

Type Definitions

Cmd 	 [Unstable]

DecodeInlinedItem 	 [Unstable]