Module rustc_trans::middle::const_eval

Unstable (rustc_private)

Structs

ConstEvalErr 	 [Unstable]

Enums

ConstVal 	 [Unstable]

ErrKind 	 [Unstable]

EvalHint 	 [Unstable]
Hint to determine how to evaluate constant expressions which might not be type-checked.

IntTy 	 [Unstable]

UintTy 	 [Unstable]

Functions

compare_const_vals 	 [Unstable]

compare_lit_exprs 	 [Unstable]

const_expr_to_pat 	 [Unstable]

const_int_checked_add 	 [Unstable]

const_int_checked_div 	 [Unstable]

const_int_checked_mul 	 [Unstable]

const_int_checked_neg 	 [Unstable]

const_int_checked_rem 	 [Unstable]

const_int_checked_shl 	 [Unstable]

const_int_checked_shl_via_uint 	 [Unstable]

const_int_checked_shr 	 [Unstable]

const_int_checked_shr_via_uint 	 [Unstable]

const_int_checked_sub 	 [Unstable]

const_uint_checked_add 	 [Unstable]

const_uint_checked_div 	 [Unstable]

const_uint_checked_mul 	 [Unstable]

const_uint_checked_neg 	 [Unstable]

const_uint_checked_rem 	 [Unstable]

const_uint_checked_shl 	 [Unstable]

const_uint_checked_shl_via_int 	 [Unstable]

const_uint_checked_shr 	 [Unstable]

const_uint_checked_shr_via_int 	 [Unstable]

const_uint_checked_sub 	 [Unstable]

eval_const_expr 	 [Unstable]

eval_const_expr_partial 	 [Unstable]
Evaluate a constant expression in a context where the expression isn't guaranteed to be evaluatable. ty_hint is usually ExprTypeChecked, but a few places need to evaluate constants during type-checking, like computing the length of an array. (See also the FIXME above EvalHint.)

lookup_const_by_id 	 [Unstable]

def_id is the id of the constant.

maybe_ref_id is the id of the expr referencing the constant.

param_substs is the monomorphization substitution for the expression.

lookup_const_fn_by_id 	 [Unstable]

Type Definitions

CastResult 	 [Unstable]

EvalResult 	 [Unstable]

FnArgMap 	 [Unstable]