Struct rustc_trans::middle::cstore::ExternCrate

pub struct ExternCrate {
    pub def_id: DefId,
    pub span: Span,
    pub direct: bool,
    pub path_len: usize,
}

Unstable (rustc_private)

Fields
def_id: DefId
Unstable (rustc_private)

def_id of an extern crate in the current crate that caused this crate to be loaded; note that there could be multiple such ids

span: Span
Unstable (rustc_private)

span of the extern crate that caused this to be loaded

direct: bool
Unstable (rustc_private)

If true, then this crate is the crate named by the extern crate referenced above. If false, then this crate is a dep of the crate.

path_len: usize
Unstable (rustc_private)

Number of links to reach the extern crate def_id declaration; used to select the extern crate with the shortest path