Struct rustc_trans::middle::mem_categorization::cmt_

pub struct cmt_<'tcx> {
    pub id: NodeId,
    pub span: Span,
    pub cat: Categorization<'tcx>,
    pub mutbl: MutabilityCategory,
    pub ty: &'tcx TyS<'tcx>,
    pub note: Note,
}

🔬 This is a nightly-only experimental API. (rustc_private)

Fields
id: NodeId

🔬 This is a nightly-only experimental API. (rustc_private)

span: Span

🔬 This is a nightly-only experimental API. (rustc_private)

cat: Categorization<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

mutbl: MutabilityCategory

🔬 This is a nightly-only experimental API. (rustc_private)

ty: &'tcx TyS<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

note: Note

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'tcx> cmt_<'tcx>

fn get_field(&self, name: Symbol) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

fn get_field_name(&self) -> Option<Symbol>

🔬 This is a nightly-only experimental API. (rustc_private)

fn get_arg_if_immutable(&self, map: &Map) -> Option<NodeId>

🔬 This is a nightly-only experimental API. (rustc_private)

impl<'tcx> cmt_<'tcx>

fn guarantor(&self) -> Rc<cmt_<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

Returns self after stripping away any derefs or interior content. The return value is basically the cmt which determines how long the value in self remains live.

fn freely_aliasable(&self) -> Aliasability

🔬 This is a nightly-only experimental API. (rustc_private)

Returns FreelyAliasable(_) if this lvalue represents a freely aliasable pointer type.

fn upvar(&self) -> Option<Rc<cmt_<'tcx>>>

🔬 This is a nightly-only experimental API. (rustc_private)

fn descriptive_string(&self, tcx: TyCtxt) -> String

🔬 This is a nightly-only experimental API. (rustc_private)