Struct rustc_trans::middle::region::RegionMaps

pub struct RegionMaps {
    // some fields omitted
}

Unstable (rustc_private)

The region maps encode information about region relationships.

Methods

impl RegionMaps

fn bogus_code_extent(&self, e: CodeExtentData) -> CodeExtent

Unstable (rustc_private)

create a bogus code extent for the regions in astencode types. Nobody really cares about the contents of these.

fn lookup_code_extent(&self, e: CodeExtentData) -> CodeExtent

fn node_extent(&self, n: u32) -> CodeExtent

fn item_extent(&self, n: u32) -> CodeExtent

fn call_site_extent(&self, fn_id: u32, body_id: u32) -> CodeExtent

fn opt_destruction_extent(&self, n: u32) -> Option<CodeExtent>

fn intern_code_extent(&self, e: CodeExtentData, parent: CodeExtent) -> CodeExtent

fn intern_node(&self, n: u32, parent: CodeExtent) -> CodeExtent

fn code_extent_data(&self, e: CodeExtent) -> CodeExtentData

fn each_encl_scope<E>(&self, e: E) where E: FnMut(&CodeExtent, &CodeExtent) -> ()

fn each_var_scope<E>(&self, e: E) where E: FnMut(&u32, &CodeExtent) -> ()

fn each_rvalue_scope<E>(&self, e: E) where E: FnMut(&u32, &CodeExtent) -> ()

fn opt_encl_scope(&self, id: CodeExtent) -> Option<CodeExtent>

Unstable (rustc_private)

Returns the narrowest scope that encloses id, if any.

fn encl_scope(&self, id: CodeExtent) -> CodeExtent

Unstable (rustc_private)

Returns the narrowest scope that encloses id, if any.

fn var_scope(&self, var_id: u32) -> CodeExtent

Unstable (rustc_private)

Returns the lifetime of the local variable var_id

fn temporary_scope(&self, expr_id: u32) -> Option<CodeExtent>

Unstable (rustc_private)

Returns the scope when temp created by expr_id will be cleaned up

fn var_region(&self, id: u32) -> Region

Unstable (rustc_private)

Returns the lifetime of the variable id.

fn scopes_intersect(&self, scope1: CodeExtent, scope2: CodeExtent) -> bool

fn is_subscope_of(&self, subscope: CodeExtent, superscope: CodeExtent) -> bool

Unstable (rustc_private)

Returns true if subscope is equal to or is lexically nested inside superscope and false otherwise.

fn nearest_common_ancestor(&self, scope_a: CodeExtent, scope_b: CodeExtent) -> CodeExtent

Unstable (rustc_private)

Finds the nearest common ancestor (if any) of two scopes. That is, finds the smallest scope which is greater than or equal to both scope_a and scope_b.