Enum rustc_trans::middle::traits::ObligationCauseCode

pub enum ObligationCauseCode<'tcx> {
    MiscObligation,
    RFC1214(Rc<ObligationCauseCode<'tcx>>),
    SliceOrArrayElem,
    ProjectionWf(ProjectionTy<'tcx>),
    ItemObligation(DefId),
    ReferenceOutlivesReferent(&'tcx TyS<'tcx>),
    ObjectCastObligation(&'tcx TyS<'tcx>),
    AssignmentLhsSized,
    StructInitializerSized,
    VariableType(u32),
    ReturnType,
    RepeatVec,
    ClosureCapture(u32, Span, BuiltinBound),
    FieldSized,
    SharedStatic,
    BuiltinDerivedObligation(DerivedObligationCause<'tcx>),
    ImplDerivedObligation(DerivedObligationCause<'tcx>),
    CompareImplMethodObligation,
}

Unstable (rustc_private)

Variants

MiscObligation
Unstable (rustc_private)

Not well classified or should be obvious from span.

RFC1214
Unstable (rustc_private)

Obligation that triggers warning until RFC 1214 is fully in place.

SliceOrArrayElem
Unstable (rustc_private)

This is the trait reference from the given projection

ProjectionWf
Unstable (rustc_private)

This is the trait reference from the given projection

ItemObligation
Unstable (rustc_private)

In an impl of trait X for type Y, type Y must also implement all supertraits of X.

ReferenceOutlivesReferent
Unstable (rustc_private)

A type like &'a T is WF only if T: 'a.

ObjectCastObligation
Unstable (rustc_private)

Obligation incurred due to an object cast.

AssignmentLhsSized
Unstable (rustc_private)

Various cases where expressions must be sized/copy/etc:

StructInitializerSized
Unstable (rustc_private)

VariableType
Unstable (rustc_private)

ReturnType
Unstable (rustc_private)

RepeatVec
Unstable (rustc_private)

ClosureCapture
Unstable (rustc_private)

FieldSized
Unstable (rustc_private)

SharedStatic
Unstable (rustc_private)

BuiltinDerivedObligation
Unstable (rustc_private)

ImplDerivedObligation
Unstable (rustc_private)

CompareImplMethodObligation
Unstable (rustc_private)

Methods

impl<'tcx> ObligationCauseCode<'tcx>

fn is_rfc1214(&self) -> bool

Unstable (rustc_private)