Enum rustc_trans::middle::ty::FragmentInfo

pub enum FragmentInfo {
    pub Moved {
        pub var: u32,
        pub move_expr: u32,
    },
    pub Assigned {
        pub var: u32,
        pub assign_expr: u32,
        pub assignee_id: u32,
    },
}

Unstable

Describes the fragment-state associated with a NodeId.

Currently only unfragmented paths have entries in the table, but longer-term this enum is expected to expand to also include data for fragmented paths.

Variants

Moved
Unstable

Fields

var
Unstable

move_expr
Unstable

Assigned
Unstable

Fields

var
Unstable

assign_expr
Unstable

assignee_id
Unstable