Enum rustc_trans::middle::ty::ObjectLifetimeDefault

pub enum ObjectLifetimeDefault {
    Ambiguous,
    BaseDefault,
    Specific(Region),
}

Unstable

Default region to use for the bound of objects that are supplied as the value for this type parameter. This is derived from T:'a annotations appearing in the type definition. If this is None, then the default is inherited from the surrounding context. See RFC #599 for details.

Variants

Ambiguous
Unstable

Require an explicit annotation. Occurs when multiple T:'a constraints are found.

BaseDefault
Unstable

Use the base default, typically 'static, but in a fn body it is a fresh variable

Specific
Unstable

Use the given region as the default.