Struct rustc_trans::middle::ty::AdtDefData

pub struct AdtDefData<'tcx, 'container> where 'container: 'tcx {
    pub did: DefId,
    pub variants: Vec<VariantDefData<'tcx, 'container>>,
    // some fields omitted
}

Unstable

The definition of an abstract data type - a struct or enum.

These are all interned (by intern_adt_def) into the adt_defs table.

Because of the possibility of nested tcx-s, this type needs 2 lifetimes: the traditional variant lifetime ('tcx) bounding the lifetime of the inner types is of course necessary. However, it is not sufficient - types from a child tcx must not be leaked into the master tcx by being stored in an AdtDefData.

The 'container lifetime ensures that by outliving the container tcx and preventing shorter-lived types from being inserted. When write access is not needed, the 'container lifetime can be erased to 'static, which can be done by the AdtDef wrapper.

Fields

did
Unstable

variants
Unstable

Methods

impl<'tcx, 'container> AdtDefData<'tcx, 'container>

fn adt_kind(&self) -> AdtKind

Unstable

Returns the kind of the ADT - Struct or Enum.

fn is_dtorck(&'tcx self, tcx: &ctxt<'tcx>) -> bool

Unstable

Returns whether this is a dtorck type. If this returns true, this type being safe for destruction requires it to be alive; Otherwise, only the contents are required to be.

fn is_fundamental(&self) -> bool

Unstable

Returns whether this type is #[fundamental] for the purposes of coherence checking.

fn is_simd(&self) -> bool

Unstable

fn is_phantom_data(&self) -> bool

Unstable

Returns true if this is PhantomData.

fn has_dtor(&self, tcx: &ctxt<'tcx>) -> bool

Unstable

Returns whether this type has a destructor.

fn struct_variant(&self) -> &VariantDefData<'tcx, 'container>

Unstable

Asserts this is a struct and returns the struct's unique variant.

fn type_scheme(&self, tcx: &ctxt<'tcx>) -> TypeScheme<'tcx>

Unstable

fn predicates(&self, tcx: &ctxt<'tcx>) -> GenericPredicates<'tcx>

Unstable

fn all_fields(&self) -> FlatMap<Iter<VariantDefData<'tcx, 'container>>, Iter<FieldDefData<'tcx, 'container>>, fn(&'s VariantDefData<'tcx, 'container>) -> Iter<'s, FieldDefData<'tcx, 'container>>>

Unstable

Returns an iterator over all fields contained by this ADT.

fn is_empty(&self) -> bool

Unstable

fn is_univariant(&self) -> bool

Unstable

fn is_payloadfree(&self) -> bool

Unstable

fn variant_with_id(&self, vid: DefId) -> &VariantDefData<'tcx, 'container>

Unstable

fn variant_of_def(&self, def: Def) -> &VariantDefData<'tcx, 'container>

Unstable