Struct rustc_trans::middle::ty::CrateAnalysis

pub struct CrateAnalysis<'a> {
    pub export_map: HashMap<u32, Vec<Export>, DefaultState<FnvHasher>>,
    pub access_levels: AccessLevels<u32>,
    pub reachable: HashSet<u32, DefaultState<FnvHasher>>,
    pub name: &'a str,
    pub glob_map: Option<HashMap<u32, HashSet<Name, RandomState>, RandomState>>,
}

Unstable (rustc_private)

The complete set of all analyses described in this module. This is produced by the driver and fed to trans and later passes.

Fields

export_map
Unstable (rustc_private)

access_levels
Unstable (rustc_private)

reachable
Unstable (rustc_private)

name
Unstable (rustc_private)

glob_map
Unstable (rustc_private)