Struct rustc_trans::middle::ty::ExistentialBounds

pub struct ExistentialBounds<'tcx> {
    pub region_bound: Region,
    pub builtin_bounds: BuiltinBounds,
    pub projection_bounds: Vec<Binder<ProjectionPredicate<'tcx>>>,
}

Unstable

Bounds suitable for an existentially quantified type parameter such as those that appear in object types or closure types.

Fields

region_bound
Unstable

builtin_bounds
Unstable

projection_bounds
Unstable