Struct rustc_trans::middle::ty::InstantiatedPredicates

pub struct InstantiatedPredicates<'tcx> {
    pub predicates: VecPerParamSpace<Predicate<'tcx>>,
}

Unstable (rustc_private)

Represents the bounds declared on a particular set of type parameters. Should eventually be generalized into a flag list of where clauses. You can obtain a InstantiatedPredicates list from a GenericPredicates by using the instantiate method. Note that this method reflects an important semantic invariant of InstantiatedPredicates: while the GenericPredicates are expressed in terms of the bound type parameters of the impl/trait/whatever, an InstantiatedPredicates instance represented a set of bounds for some particular instantiation, meaning that the generic parameters have been substituted with their values.

Example:

struct Foo<T,U:Bar<T>> { ... }

Here, the GenericPredicates for Foo would contain a list of bounds like [[], [U:Bar<T>]]. Now if there were some particular reference like Foo<isize,usize>, then the InstantiatedPredicates would be [[], [usize:Bar<isize>]].

Fields

predicates
Unstable (rustc_private)

Methods

impl<'tcx> InstantiatedPredicates<'tcx>

fn empty() -> InstantiatedPredicates<'tcx>

Unstable (rustc_private)

fn has_escaping_regions(&self) -> bool

Unstable (rustc_private)

fn is_empty(&self) -> bool

Unstable (rustc_private)