Trait rustc_trans::middle::ty::TypeFoldable

pub trait TypeFoldable<'tcx>: Clone + Debug {
    fn super_fold_with<F>(&self, folder: &mut F) -> Self where F: TypeFolder<'tcx>;
    fn super_visit_with<V>(&self, visitor: &mut V) -> bool where V: TypeVisitor<'tcx>;

    fn fold_with<F>(&self, folder: &mut F) -> Self where F: TypeFolder<'tcx> { ... }
    fn visit_with<V>(&self, visitor: &mut V) -> bool where V: TypeVisitor<'tcx> { ... }
    fn has_regions_escaping_depth(&self, depth: u32) -> bool { ... }
    fn has_escaping_regions(&self) -> bool { ... }
    fn has_type_flags(&self, flags: TypeFlags) -> bool { ... }
    fn has_projection_types(&self) -> bool { ... }
    fn references_error(&self) -> bool { ... }
    fn has_param_types(&self) -> bool { ... }
    fn has_self_ty(&self) -> bool { ... }
    fn has_infer_types(&self) -> bool { ... }
    fn needs_infer(&self) -> bool { ... }
    fn needs_subst(&self) -> bool { ... }
    fn has_closure_types(&self) -> bool { ... }
    fn has_erasable_regions(&self) -> bool { ... }
    fn is_global(&self) -> bool { ... }
}

Unstable (rustc_private)

The TypeFoldable trait is implemented for every type that can be folded. Basically, every type that has a corresponding method in TypeFolder.

Required Methods

fn super_fold_with<F>(&self, folder: &mut F) -> Self where F: TypeFolder<'tcx>

Unstable (rustc_private)

fn super_visit_with<V>(&self, visitor: &mut V) -> bool where V: TypeVisitor<'tcx>

Unstable (rustc_private)

Provided Methods

fn fold_with<F>(&self, folder: &mut F) -> Self where F: TypeFolder<'tcx>

Unstable (rustc_private)

fn visit_with<V>(&self, visitor: &mut V) -> bool where V: TypeVisitor<'tcx>

Unstable (rustc_private)

fn has_regions_escaping_depth(&self, depth: u32) -> bool

Unstable (rustc_private)

fn has_escaping_regions(&self) -> bool

Unstable (rustc_private)

fn has_type_flags(&self, flags: TypeFlags) -> bool

Unstable (rustc_private)

fn has_projection_types(&self) -> bool

Unstable (rustc_private)

fn references_error(&self) -> bool

Unstable (rustc_private)

fn has_param_types(&self) -> bool

Unstable (rustc_private)

fn has_self_ty(&self) -> bool

Unstable (rustc_private)

fn has_infer_types(&self) -> bool

Unstable (rustc_private)

fn needs_infer(&self) -> bool

Unstable (rustc_private)

fn needs_subst(&self) -> bool

Unstable (rustc_private)

fn has_closure_types(&self) -> bool

Unstable (rustc_private)

fn has_erasable_regions(&self) -> bool

Unstable (rustc_private)

fn is_global(&self) -> bool

Unstable (rustc_private)

Indicates whether this value references only 'global' types/lifetimes that are the same regardless of what fn we are in. This is used for caching. Errs on the side of returning false.