Module rustc_trans::save

Unstable (rustc_private)

Modules

recorder 	 [Unstable]

span_utils 	 [Unstable]

Structs

CrateData 	 [Unstable]

EnumData 	 [Unstable]
Data for enum declarations.

FunctionCallData 	 [Unstable]
Data about a function call.

FunctionData 	 [Unstable]
Data for all kinds of functions and methods.

ImplData 	 [Unstable]

MethodCallData 	 [Unstable]
Data about a method call.

ModData 	 [Unstable]
Data for modules.

ModRefData 	 [Unstable]
Data for a reference to a module.

SaveContext 	 [Unstable]

TypeRefData 	 [Unstable]
Data for a reference to a type or trait.

VariableData 	 [Unstable]
Data for local and global variables (consts and statics).

VariableRefData 	 [Unstable]
Data for the use of some item (e.g., the use of a local variable, which will refer to that variables declaration (by ref_id)).

Enums

Data 	 [Unstable]
Data for any entity in the Rust language. The actual data contained varied with the kind of entity being queried. See the nested structs for details.

Functions

generated_code 	 [Unstable]

process_crate 	 [Unstable]