Enum rustc_trans::save::Data

pub enum Data {
    FunctionData(FunctionData),
    VariableData(VariableData),
    ModData(ModData),
    EnumData(EnumData),
    ImplData(ImplData),
    VariableRefData(VariableRefData),
    TypeRefData(TypeRefData),
    ModRefData(ModRefData),
    FunctionCallData(FunctionCallData),
    MethodCallData(MethodCallData),
}

Unstable

Data for any entity in the Rust language. The actual data contained varied with the kind of entity being queried. See the nested structs for details.

Variants

FunctionData
Unstable

Data for all kinds of functions and methods.

VariableData
Unstable

Data for local and global variables (consts and statics), and fields.

ModData
Unstable

Data for modules.

EnumData
Unstable

Data for Enums.

ImplData
Unstable

Data for impls.

VariableRefData
Unstable

Data for the use of some variable (e.g., the use of a local variable, which will refere to that variables declaration).

TypeRefData
Unstable

Data for a reference to a type or trait.

ModRefData
Unstable

Data for a reference to a module.

FunctionCallData
Unstable

Data about a function call.

MethodCallData
Unstable

Data about a method call.