Module rustc_trans::util::nodemap

Unstable

An efficient hash map for node IDs

Structs

FnvHasher 	 [Unstable]
A speedy hash algorithm for node ids and def ids. The hashmap in libcollections by default uses SipHash which isn't quite as speedy as we want. In the compiler we're not really worried about DOS attempts, so we just default to a non-cryptographic hash.

Functions

DefIdMap 	 [Unstable]

DefIdSet 	 [Unstable]

FnvHashMap 	 [Unstable]

FnvHashSet 	 [Unstable]

NodeMap 	 [Unstable]

NodeSet 	 [Unstable]

Type Definitions

DefIdMap 	 [Unstable]

DefIdSet 	 [Unstable]

FnvHashMap 	 [Unstable]

FnvHashSet 	 [Unstable]

NodeMap 	 [Unstable]

NodeSet 	 [Unstable]