Enum rustc_trans::middle::traits::ObligationCauseCodeUnstable

pub enum ObligationCauseCode<'tcx> {
    MiscObligation,
    ItemObligation(DefId),
    ObjectCastObligation(&'tcx TyS<'tcx>),
    AssignmentLhsSized,
    StructInitializerSized,
    VariableType(u32),
    ReturnType,
    RepeatVec,
    ClosureCapture(u32, Span, BuiltinBound),
    FieldSized,
    ObjectSized,
    SharedStatic,
    BuiltinDerivedObligation(DerivedObligationCause<'tcx>),
    ImplDerivedObligation(DerivedObligationCause<'tcx>),
    CompareImplMethodObligation,
}

Variants

MiscObligation

Not well classified or should be obvious from span.

ItemObligation

In an impl of trait X for type Y, type Y must also implement all supertraits of X.

ObjectCastObligation

Obligation incurred due to an object cast.

AssignmentLhsSized

Various cases where expressions must be sized/copy/etc:

StructInitializerSized

VariableType

ReturnType

RepeatVec

ClosureCapture

FieldSized

ObjectSized

SharedStatic

BuiltinDerivedObligation

ImplDerivedObligation

CompareImplMethodObligation