Enum rustc_trans::middle::ty::ObjectLifetimeDefaultUnstable

pub enum ObjectLifetimeDefault {
    Ambiguous,
    Specific(Region),
}

Default region to use for the bound of objects that are supplied as the value for this type parameter. This is derived from T:'a annotations appearing in the type definition. If this is None, then the default is inherited from the surrounding context. See RFC #599 for details.

Variants

Ambiguous

Require an explicit annotation. Occurs when multiple T:'a constraints are found.

Specific

Use the given region as the default.